fix(api): use int64 for BGP RemoteASN to avoid int32 CEL format rejection
Kubernetes API servers now strictly enforce the OpenAPI 'format' on
numeric fields (int32/int64 are the only formats OpenAPI defines —
there is no unsigned format). controller-gen maps Go's uint32 to
'format: int32', which is signed and tops out at 2147483647.

BGPPeer.RemoteASN and the legacy BGPPeeringSpec.RemoteASN were declared
as uint32, so any 4-byte ASN above 2147483647 (e.g. 4200011028, a
common value in the RFC 6996 private ASN range) got rejected by the
API server when applying the generated NodeNetworkConfig:

  NodeNetworkConfig.network.t-caas.telekom.com "<node>" is invalid:
  spec.fabricVRFs.<vrf>.bgpPeers[0].remoteAsn: Invalid value: "":
  Checked value must be of type integer with format int32

Switch both fields to int64 (format: int64, safely covering the full
uint32 range) with explicit Minimum=1/Maximum=4294967295 markers,
matching the already-correct BGPPeering.spec.workloadAS *int64
pattern. CRDs regenerated via controller-gen; call sites and tests
updated accordingly.

+++ api/v1alpha1/bgppeering_types.go
@@ -40,8 +40,15 @@ type BGPPeeringSpec struct {
 	// Loopback is the loopback interface used for the BGP peering
 	LoopbackPeer *BGPPeeringLoopback `json:"loopbackPeer,omitempty"`
 
-	// RemoteASN is the ASN of the remote BGP peer
-	RemoteASN uint32 `json:"remoteASN"`
+	// RemoteASN is the ASN of the remote BGP peer. Uses asplain notation;
+	// 4-byte ASNs (up to 4294967295) are common and exceed the signed int32
+	// range, so this is int64 (format: int64) rather than uint32/int32 —
+	// controller-gen maps uint32 to the OpenAPI "int32" format, whose CEL
+	// format-check enforces the *signed* int32 range (max 2147483647) and
+	// would reject legitimate large 4-byte ASNs.
+	// +kubebuilder:validation:Minimum=1
+	// +kubebuilder:validation:Maximum=4294967295
+	RemoteASN int64 `json:"remoteASN"`
 	// EnableBFD is the flag to enable BFD for the BGP peering
 	EnableBFD bool `json:"enableBFD"`
